dropdown categories seeding associated and couple bug fixes
