+++ client/components/SelectedCategory.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default class SelectedCategory extends React.Component {
+  componentDidMount() {}
+}
