+++ script/seedGenerator.js
@@ -1,62 +1,62 @@
 'use strict'
 
 var casual = require('casual')
-var totalUsersToFake = 50
-
-const servicesList = [
-  'HTML Tutor',
-  'CSS Tutor',
-  'Lawn Mowing',
-  'Modeling',
-  'Picture Taking',
-  'Personal Training',
-  'Financial Analysis',
-  'Doing The Dishes',
-  'Taking Out The Crash',
-  'SAT Tutoring',
-  'ACT Tutoring',
-  'Piano Lessons',
-  'Jazz Lessons',
-  'Window Cleaning',
-  'House Keeping',
-  'Hair-Cutting',
-  'Tattoo Design',
-  'Caricature Drawing',
-  'Personal Chef',
-  'Personal Bartender',
-  'Personal Stylist',
-  'Basketball Lessons',
-  'Interior Design',
-  'Meditation Practice',
-  'Help Doing Taxes',
-  'Personal Nutritionist',
-  'Music Production',
-  'Guitar Lessons',
-  'Cantonese Translator',
-  'Nanny',
-  'Physical Therapy',
-  'Psychological Therapy',
-  'Nail Technician',
-  'Hair Stylist',
-  'Make-up',
-  'House Makeover',
-  'Japanese Tutor',
-  'Life Makeover',
-  'Javascript Tutoring',
-  'Photoshopping Any Image',
-  'SQL Tutoring',
-  'Python 3 Tutoring',
-  'Java Tutoring',
-  'Social Skill Tutor',
-  'Express.js Tutoring',
-  'React.js Tutoring',
-  'Node Tutoring',
-  'Redux.js Tutoring',
-  'Vocal Lessons',
-  'Personal Hype Man',
-  'Algebra Tutoring',
-  'Travel Guide'
-]
+var totalUsersToFake = 200
+const servicesWithCategory = {
+  'HTML Tutor': 1,
+  'CSS Tutor': 1,
+  'SQL Tutoring': 1,
+  'Python 3 Tutoring': 1,
+  'Java Tutoring': 1,
+  'Japanese Tutor': 1,
+  'Javascript Tutoring': 1,
+  'Algebra Tutoring': 1,
+  'SAT Tutoring': 1,
+  'ACT Tutoring': 1,
+  'Express.js Tutoring': 1,
+  'React.js Tutoring': 1,
+  'Cantonese Translator': 1,
+  'Node Tutoring': 1,
+  'Redux.js Tutoring': 1,
+  'Vocal Lessons': 2,
+  'Piano Lessons': 2,
+  'Jazz Lessons': 2,
+  'Guitar Lessons': 2,
+  'Music Production': 2,
+  'Hair-Cutting': 3,
+  'Tattoo Design': 3,
+  'Nail Technician': 3,
+  'Hair Stylist': 3,
+  'Make-up': 3,
+  'Personal Bartender': 3,
+  'Personal Stylist': 3,
+  Modeling: 3,
+  'Physical Therapy': 4,
+  'Psychological Therapy': 4,
+  'House Makeover': 5,
+  'Lawn Mowing': 5,
+  'Taking Out The Trash': 5,
+  'Window Cleaning': 5,
+  'House Keeping': 5,
+  'Doing The Dishes': 5,
+  Nanny: 5,
+  'Personal Chef': 5,
+  'Interior Design': 5,
+  'Basketball Lessons': 6,
+  'Personal Training': 6,
+  'Picture Taking': 7,
+  'Caricature Drawing': 7,
+  'Financial Analysis': 8,
+  'Photoshopping Any Image': 8,
+  'Meditation Practice': 9,
+  'Help Doing Taxes': 9,
+  'Personal Nutritionist': 9,
+  'Social Skill Tutor': 9,
+  'Personal Hype Man': 9,
+  'Life Makeover': 9,
+  'Travel Guide': 9
+}
+const servicesList = Object.keys(servicesWithCategory)
 
 const reviewList = [
   'Trash',
@@ -135,12 +135,227 @@ const imageUrls = [
   'https://ca.slack-edge.com/T024FPYBQ-UHF7SFX1P-71576b1c3c80-512'
 ]
 
+const interestsList = [
+  'Baseball',
+  'Cooking',
+  'Pottery',
+  'Comedy Clubs',
+  'Skiing',
+  'Talking',
+  'Myspace',
+  'TikTok',
+  'Owl Watching',
+  'Ice Cream',
+  'Playing I Spy',
+  'Apple Products',
+  'Golf',
+  'Country Clubbing',
+  'Netflix',
+  'Mathas Vineyard',
+  'Reading',
+  'Going to The Zoo',
+  'Aquariums',
+  'DOTA 2',
+  'Bragging',
+  'Shoplifting',
+  'Meditation',
+  'Watches',
+  'Laundry',
+  '2048',
+  'Jewelry',
+  'Joker Movie',
+  'Elon Musk',
+  'Highschool Musical',
+  'Dinosaurs',
+  'Bird Watching',
+  'The Incredibles 1',
+  'Nintendo Switch',
+  'Super Smash Bros',
+  'Flexing',
+  'Machine Learning',
+  'Apple Maps',
+  'Boggle With Friends',
+  'Scrabble With Friends',
+  'Farmville: Darkness Rising',
+  'Sims 3: Unrated',
+  'Playing the Piano',
+  'Teletubbies',
+  'Boo-Bah',
+  'Barney and Friends',
+  'Rob Fox',
+  'Hanging out with Eric Yang',
+  'Running Marathons',
+  'Avengers On Ice',
+  'Traveling',
+  'AlgoExpert',
+  'LeetCode',
+  'Transformers 2: The Musical',
+  'Gambling',
+  'Winnie The Pooh: Darkness Rising',
+  'Squash',
+  'Bocceball',
+  'Volleyball',
+  'Polo',
+  "Where's Waldo?",
+  'Jim Carrey',
+  'Pet Ownership',
+  'Sailing',
+  'Greek Fodd',
+  'Drinking Prosecco',
+  'Vineyard Vines',
+  'Patagonia',
+  'Personal Training',
+  'Snowball Fighting',
+  'Pescatarianism',
+  'Veganism',
+  'Fast Food',
+  'Bitcoin',
+  'Shoe Shopping',
+  'Cracking Jokes',
+  '5Gum',
+  'Germophobia',
+  'Razor Scooters',
+  'Competitve Tickling',
+  'James Bond',
+  'Club Getaway',
+  'Club Med',
+  'Getting Spray Tans',
+  'Getting my nails done',
+  'Giving Fist Bumps',
+  'Cheech and Chong',
+  'Cardio',
+  'Tarantino',
+  'Eternal Sunshine of a Spotless Mind',
+  'Pasta with butter',
+  'Glassblowing',
+  'Being kind',
+  'Painting Large Paintings',
+  'Drawing',
+  'Living In The Darkness(Shadows)',
+  'Teleworking',
+  'The American Flag',
+  'Philanthropy',
+  'Acrobatics',
+  'Building Furniture',
+  'Learning',
+  'Shopping',
+  'Lock Picking',
+  'Eating a lot',
+  'Career Development',
+  'Cleaning the house',
+  'Washing dishes',
+  'Video Games',
+  'Basketball',
+  'Anime',
+  'Comics',
+  'Film',
+  'Painting',
+  'Dance',
+  'Martial Arts',
+  'Computer Programming',
+  'Scuba Diving',
+  'Biking',
+  'Rock Climbing',
+  'Hiking',
+  'Dodgeball',
+  'Kickball',
+  'Kickboxing',
+  'MMA',
+  'Kung Fu-Movies',
+  'Psychology',
+  'Comedy',
+  'History',
+  'WW2',
+  'The American Civil War',
+  'Fashion',
+  'Card Tricks',
+  'Weightlifting',
+  'Board Games',
+  'Yu-Gi-Oh',
+  'Pokemon',
+  'Drowning My Demons',
+  'Staring At the Ceiling',
+  'Pretending',
+  'Projecting',
+  'Having friends',
+  'Water Polo',
+  'Swimming',
+  'Watching Movies',
+  'Getting tea',
+  'Ramen',
+  'Enlightenment',
+  'Yoga',
+  'Alternative Music',
+  'Deep Conversations',
+  'Doing Nails',
+  'Getting my hair done',
+  'Gossiping',
+  'Vibing',
+  'White Noise',
+  'Chilling',
+  'Hanging out',
+  'Listening to music',
+  'Trumbone',
+  'Violin',
+  'Competitive Sky Diving',
+  'Listening to Drake',
+  'Talking about Money on my bluetooth in financial district',
+  'Pasta with sauce',
+  'Being deep',
+  'Talking about space and aliens and other deep topics',
+  'Crying to Drake',
+  'Algos',
+  'Ping Pong Tournaments',
+  'Gaming',
+  'Dungeons And Dragons',
+  'Crying in The Shower',
+  'Magic Tricks',
+  'Magic The Gathering',
+  'Body Building',
+  'Science Experiments',
+  'Learning languages',
+  'Boxing',
+  'Choir',
+  'Soccer',
+  'Checking stocks and having no insight',
+  'Reading about Start-Ups',
+  'Debating',
+  'Voting',
+  'Prank Calling',
+  'Making memes',
+  'Hanging out with my stuffed animals',
+  'The Illuminati',
+  'Cults',
+  'Secret Societies',
+  'Drinking coffee',
+  'Making coffee',
+  'Selling used goods',
+  'Politics',
+  'Banking',
+  'Doing my taxes',
+  'Commuting',
+  'Urban exploring',
+  'Baking',
+  'Fishing',
+  'Competitve Fishing',
+  'Bowling',
+  'Streetwear',
+  'Searching for meaning',
+  'Meetups',
+  'Networking',
+  'LinkedIn',
+  'Practicing my pitch',
+  'Working on my resume',
+  'Being productive at work',
+  'Non-matching socks'
+]
+
 const numberOfServices = 154
 function generateReviews() {
   const reviews = []
   for (let i = 0; i < 1000; i++) {
     reviews.push({
-      rating: Math.ceil(Math.random() * 5) * 1.0,
+      rating: 2 + Math.ceil(Math.random() * 3) * 1.0,
       comment: reviewList[Math.floor(Math.random() * reviewList.length)],
       serviceId: Math.ceil(Math.random() * numberOfServices),
       userId: Math.ceil(Math.random() * (totalUsersToFake + 4))
@@ -172,14 +387,15 @@ function generateUsers(num) {
 
 function generateServices(num) {
   const services = []
+  const name = servicesList[Math.floor(Math.random() * servicesList.length)]
   for (let i = 5; i < num * 3 + 5; i++) {
     services.push({
-      serviceCategoryId: Math.ceil(Math.random() * 7),
       userId: 1 + Math.floor(Math.random() * (num + 4)),
-      name: servicesList[Math.floor(Math.random() * servicesList.length)],
+      name: name,
       description: casual.text,
       proficiency: Math.ceil(Math.random() * 3),
-      remote: casual.coin_flip
+      remote: casual.coin_flip,
+      serviceCategoryId: servicesWithCategory[name]
       // serviceRating: Math.floor(5 * Math.random()),
       // reviewCount: 1
     })
@@ -196,5 +412,6 @@ module.exports = {
   fakeServices,
   generateUsers,
   generateServices,
-  fakeReviews
+  fakeReviews,
+  interestsList
 }
