+++ server/algorithm.js
@@ -1,219 +1,5 @@
 const util = require('util')
-
-const interestsList = [
-  'Baseball',
-  'Cooking',
-  'Pottery',
-  'Comedy Clubs',
-  'Skiing',
-  'Talking',
-  'Myspace',
-  'TikTok',
-  'Owl Watching',
-  'Ice Cream',
-  'Playing I Spy',
-  'Apple Products',
-  'Golf',
-  'Country Clubbing',
-  'Netflix',
-  'Mathas Vineyard',
-  'Reading',
-  'Going to The Zoo',
-  'Aquariums',
-  'DOTA 2',
-  'Bragging',
-  'Shoplifting',
-  'Meditation',
-  'Watches',
-  'Laundry',
-  '2048',
-  'Jewelry',
-  'Joker Movie',
-  'Elon Musk',
-  'Highschool Musical',
-  'Dinosaurs',
-  'Bird Watching',
-  'The Incredibles 1',
-  'Nintendo Switch',
-  'Super Smash Bros',
-  'Flexing',
-  'Machine Learning',
-  'Apple Maps',
-  'Boggle With Friends',
-  'Scrabble With Friends',
-  'Farmville: Darkness Rising',
-  'Sims 3: Unrated',
-  'Playing the Piano',
-  'Teletubbies',
-  'Boo-Bah',
-  'Barney and Friends',
-  'Rob Fox',
-  'Hanging out with Eric Yang',
-  'Running Marathons',
-  'Avengers On Ice',
-  'Traveling',
-  'AlgoExpert',
-  'LeetCode',
-  'Transformers 2: The Musical',
-  'Gambling',
-  'Winnie The Pooh: Darkness Rising',
-  'Squash',
-  'Bocceball',
-  'Volleyball',
-  'Polo',
-  "Where's Waldo?",
-  'Jim Carrey',
-  'Pet Ownership',
-  'Sailing',
-  'Greek Fodd',
-  'Drinking Prosecco',
-  'Vineyard Vines',
-  'Patagonia',
-  'Personal Training',
-  'Snowball Fighting',
-  'Pescatarianism',
-  'Veganism',
-  'Fast Food',
-  'Bitcoin',
-  'Shoe Shopping',
-  'Cracking Jokes',
-  '5Gum',
-  'Germophobia',
-  'Razor Scooters',
-  'Competitve Tickling',
-  'James Bond',
-  'Club Getaway',
-  'Club Med',
-  'Getting Spray Tans',
-  'Getting my nails done',
-  'Giving Fist Bumps',
-  'Cheech and Chong',
-  'Cardio',
-  'Tarantino',
-  'Eternal Sunshine of a Spotless Mind',
-  'Pasta with butter',
-  'Glassblowing',
-  'Being kind',
-  'Painting Large Paintings',
-  'Drawing',
-  'Living In The Darkness(Shadows)',
-  'Teleworking',
-  'The American Flag',
-  'Philanthropy',
-  'Acrobatics',
-  'Building Furniture',
-  'Learning',
-  'Shopping',
-  'Lock Picking',
-  'Eating a lot',
-  'Career Development',
-  'Cleaning the house',
-  'Washing dishes',
-  'Video Games',
-  'Basketball',
-  'Anime',
-  'Comics',
-  'Film',
-  'Painting',
-  'Dance',
-  'Martial Arts',
-  'Computer Programming',
-  'Scuba Diving',
-  'Biking',
-  'Rock Climbing',
-  'Hiking',
-  'Dodgeball',
-  'Kickball',
-  'Kickboxing',
-  'MMA',
-  'Kung Fu-Movies',
-  'Psychology',
-  'Comedy',
-  'History',
-  'WW2',
-  'The American Civil War',
-  'Fashion',
-  'Card Tricks',
-  'Weightlifting',
-  'Board Games',
-  'Yu-Gi-Oh',
-  'Pokemon',
-  'Drowning My Demons',
-  'Staring At the Ceiling',
-  'Pretending',
-  'Projecting',
-  'Having friends',
-  'Water Polo',
-  'Swimming',
-  'Watching Movies',
-  'Getting tea',
-  'Ramen',
-  'Enlightenment',
-  'Yoga',
-  'Alternative Music',
-  'Deep Conversations',
-  'Doing Nails',
-  'Getting my hair done',
-  'Gossiping',
-  'Vibing',
-  'White Noise',
-  'Chilling',
-  'Hanging out',
-  'Listening to music',
-  'Trumbone',
-  'Violin',
-  'Competitive Sky Diving',
-  'Listening to Drake',
-  'Talking about Money on my bluetooth in financial district',
-  'Being hot',
-  'Being deep',
-  'Talking about space and aliens and other deep topics',
-  'Crying to Drake',
-  'Algos',
-  'Vaping',
-  'Gaming',
-  'Dungeons And Dragons',
-  'Crying in The Shower',
-  'Magic Tricks',
-  'Magic The Gathering',
-  'Body Building',
-  'Science Experiments',
-  'Learning languages',
-  'Boxing',
-  'Choir',
-  'Soccer',
-  'Checking stocks and having no insight',
-  'Reading about Start-Ups',
-  'Debating',
-  'Voting',
-  'Prank Calling',
-  'Making memes',
-  'Hanging out with my stuffed animals',
-  'The Illuminati',
-  'Cults',
-  'Secret Societies',
-  'Drinking coffee',
-  'Making coffee',
-  'Selling used goods',
-  'Politics',
-  'Banking',
-  'Doing my taxes',
-  'Commuting',
-  'Urban exploring',
-  'Baking',
-  'Fishing',
-  'Competitve Fishing',
-  'Bowling',
-  'Streetwear',
-  'Searching for meaning',
-  'Meetups',
-  'Networking',
-  'LinkedIn',
-  'Practicing my pitch',
-  'Working on my resume',
-  'Being productive at work',
-  'Non-matching socks'
-]
+const {interestsList} = require('../script/seedGenerator')
 
 function generateRandomUsers(interests = interestsList) {
   const allUsers = []
@@ -378,7 +164,6 @@ userValues.sort((userA, userB) => {
 // });
 
 module.exports = {
-  interestsList,
   userInterestsToSet,
   getScoreFromInterests,
   getScoreFromOverallRating,
