+++ server/db/models/index.js
@@ -70,75 +70,6 @@ User.hasMany(Review)
 Exchange.belongsTo(Swap)
 Swap.hasMany(Exchange)
 
-Review.beforeCreate(async (review, options) => {
-  const reviewedService = await Service.findByPk(review.serviceId)
-  const reviewedUser = await User.findByPk(reviewedService.userId)
-
-  if (reviewedUser.reviewCount === 0) {
-    reviewedUser.overallRating = review.rating
-    reviewedUser.reviewCount = reviewedUser.reviewCount + 1
-    await reviewedUser.save()
-  } else {
-    reviewedUser.overallRating =
-      (reviewedUser.overallRating * reviewedUser.reviewCount +
-        review.rating * 1.0) /
-      (reviewedUser.reviewCount + 1)
-    reviewedUser.reviewCount = reviewedUser.reviewCount + 1
-    await reviewedUser.save()
-  }
-
-  if (reviewedService.reviewCount === 0) {
-    reviewedService.serviceRating = review.rating
-    // reviewedService.reviewCount = reviewedService.reviewCount + 1
-    await reviewedService.save()
-  } else {
-    reviewedService.serviceRating =
-      (reviewedService.serviceRating * reviewedService.reviewCount +
-        review.rating * 1.0) /
-      (reviewedService.reviewCount + 1)
-    // reviewedUser.reviewCount = reviewedUser.reviewCount + 1
-    await reviewedService.save()
-  }
-
-  await reviewedUser.increment('reviewCount', {by: 1})
-  await reviewedService.increment('reviewCount', {by: 1})
-
-  // await reviewedUser.update({
-  //   reviewCount: reviewedUser.reviewCount+1
-  // })
-
-  // await reviewedService.update({
-  //   reviewCount: reviewedService.reviewCount+1
-  // })
-  //before adding the review. calculate the new average for the specific skill
-  //rating. then calculate the new average for the overall rating. then update on
-  //the user model and the service model.
-  // await User.findByPk()
-})
-
-// Review.beforeBulkCreate(function(reviewArray) {
-//   reviewArray.forEach(async review => {
-//     try{
-
-//       const reviewedService = await Service.findByPk(review.dataValues.serviceId)
-
-//       console.log(reviewedService)
-//       const reviewedUser = await User.findByPk(reviewedService.userId)
-//       await reviewedUser.increment('reviewCount', {by: 1})
-//       await reviewedService.increment('reviewCount', {by: 1})
-//     }catch(error){
-//       console.error(error)
-//     }
-
-//     // await reviewedUser.update({
-//     //   reviewCount: reviewedUser.reviewCount + 1
-//     // })
-
-//     // await reviewedService.update({
-//     //   reviewCount: reviewedService.reviewCount + 1
-//     // })
-//   })
-// })
 /**
  * If we had any associations to make, this would be a great place to put them!
  * ex. if we had another model called BlogPost, we might say:
